+++ R/mod_addComment.R
@@ -57,7 +57,7 @@ addCommentServer <- function(id, metric_abrv, user, credentials, pkg_name) {
     
     observeEvent(input$submit_comment, {
       req(input$add_comment)
-      req("general_comment" %in% unlist(credentials$privileges[user$role], use.name = FALSE))
+      req("general_comment" %in% unlist(credentials$privileges[user$role], use.names = FALSE))
       
       comment <- trimws(input$add_comment)